requirements.txt
Django==4.1.7
Pillow==9.3.0
psycopg2==2.9.5
djangorestframework==3.14.0
